# services/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CatalogItemDetail": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "findings_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Findings Count",
            "type": "integer"
          },
          "github_org": {
            "title": "Github Org",
            "type": "string"
          },
          "github_repo": {
            "title": "Github Repo",
            "type": "string"
          },
          "github_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Github Url"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "item_metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Metadata"
          },
          "kind": {
            "enum": [
              "skill",
              "mcp_server",
              "hook",
              "plugin",
              "rules"
            ],
            "title": "Kind",
            "type": "string"
          },
          "latest_scan_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Scan At"
          },
          "latest_scan_score": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Scan Score"
          },
          "latest_scan_tier": {
            "anyOf": [
              {
                "enum": [
                  "green",
                  "yellow",
                  "orange",
                  "red",
                  "unscoped"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Scan Tier"
          },
          "popularity_score": {
            "default": 0,
            "minimum": 0.0,
            "title": "Popularity Score",
            "type": "integer"
          },
          "popularity_tier": {
            "title": "Popularity Tier",
            "type": "string"
          },
          "registries": {
            "items": {
              "type": "string"
            },
            "title": "Registries",
            "type": "array"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "sources": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Sources",
            "type": "array"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "kind",
          "display_name",
          "github_org",
          "github_repo",
          "popularity_tier",
          "updated_at"
        ],
        "title": "CatalogItemDetail",
        "type": "object"
      },
      "CatalogItemSummary": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "findings_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Findings Count",
            "type": "integer"
          },
          "github_org": {
            "title": "Github Org",
            "type": "string"
          },
          "github_repo": {
            "title": "Github Repo",
            "type": "string"
          },
          "github_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Github Url"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "skill",
              "mcp_server",
              "hook",
              "plugin",
              "rules"
            ],
            "title": "Kind",
            "type": "string"
          },
          "latest_scan_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Scan At"
          },
          "latest_scan_score": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Scan Score"
          },
          "latest_scan_tier": {
            "anyOf": [
              {
                "enum": [
                  "green",
                  "yellow",
                  "orange",
                  "red",
                  "unscoped"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Scan Tier"
          },
          "popularity_score": {
            "default": 0,
            "minimum": 0.0,
            "title": "Popularity Score",
            "type": "integer"
          },
          "popularity_tier": {
            "title": "Popularity Tier",
            "type": "string"
          },
          "registries": {
            "items": {
              "type": "string"
            },
            "title": "Registries",
            "type": "array"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "kind",
          "display_name",
          "github_org",
          "github_repo",
          "popularity_tier",
          "updated_at"
        ],
        "title": "CatalogItemSummary",
        "type": "object"
      },
      "CatalogListEnvelope": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/CatalogItemSummary"
            },
            "title": "Data",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "total_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Total Count",
            "type": "integer"
          }
        },
        "required": [
          "data"
        ],
        "title": "CatalogListEnvelope",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Every AI skill, independently scanned \u2014 public, free, Apache-2.0 trust-scoring for skills, MCP servers, hooks, and plugins.",
    "title": "SaferSkills API",
    "version": "0.0.0+foundation"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/items": {
      "get": {
        "description": "Catalog list with the per-row latest-scan join + filters.",
        "operationId": "list_items_api_v1_items_get",
        "parameters": [
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "score_min",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 100,
              "minimum": 0,
              "title": "Score Min",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "score_max",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 0,
              "title": "Score Max",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "scan_tier",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scan Tier"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "most_installed",
              "enum": [
                "most_installed",
                "recent",
                "highest_score",
                "lowest_score",
                "most_starred"
              ],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogListEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Items",
        "tags": [
          "items"
        ]
      }
    },
...
    "/api/v1/items/{slug}": {
      "get": {
        "operationId": "get_item_api_v1_items__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogItemDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Item",
        "tags": [
          "items"
        ]
      }
    },
...
  }
}